+++ pomodoroTimer/index.html
@@ -35,7 +35,8 @@
     </div>
   </div>
 
-  <script src="main.js"></script>
+  <script src="main.js" type="module"></script>
+  <script src="utils.js" type="module"></script>
 
 </body>
